[Update] 最短経路探索実装完了

+++ Phase1_new/Ex3/point.h
@@ -6,7 +6,7 @@ typedef struct {
   double y;
   int roadA;
   int roadB;
-  int cost;    // コスト
+  double cost;    // コスト
   int done;    // 訪問済みのフラグ
   int preNode; // 最短経路として選択された自身の前のノード
   int id;
